requirements.txt
discord-py-interactions==4.3.1